src/requirements.txt
discord.py == 2.6.3
websockets == 15.0.1
python-dotenv == 1.1.1
requests == 2.32.5
Pillow == 12.0.0
CairoSVG == 2.8.2